telegram_bot_requirements.txt
python-telegram-bot==20.8
aiohttp>=3.9.0
asyncio
python-dotenv>=1.0.0
